New content for Labs 6 and 7.
New content for labs 6 and 7. Improvements and fixes to previous labs. Lab 8 early revision: content same as lab 7 for now, but the CSS code has been entirely refactored for better effeciency and to make it easier to add features in the future. Also includes general fixes and improvements to the lab 8 CSS.